Split components and views
