+++ src/index.tsx
@@ -1,7 +1,7 @@
 import { StrictMode } from 'react';
 import { createRoot } from 'react-dom/client';
 import './styles/index.css';
-import App from './components/App';
+import App from './views/App';
 import { PathFindingContextProvider } from './contexts/PathFindingContext/PathFindingContextProvider';
 import { UserActionContextProvider } from './contexts/UserActionContext/UserActionContextProvider';
