+++ src/views/Dashboard/SelectEndButton.tsx
@@ -1,6 +1,6 @@
 import { useCallback } from 'react';
 import { useUserActionDispatchContext } from '../../contexts/UserActionContext';
-import Button from '../Button';
+import Button from '../../components/Button';
 
 export const SelectEndButton = ({ disabled }: { disabled: boolean }) => {
   const dispatchUserAction = useUserActionDispatchContext();
