+++ src/views/Dashboard/SelectStartButton.tsx
@@ -1,6 +1,6 @@
 import { useCallback } from 'react';
 import { useUserActionDispatchContext } from '../../contexts/UserActionContext';
-import Button from '../Button';
+import Button from '../../components/Button';
 
 export const SelectStartButton = ({ disabled }: { disabled: boolean }) => {
   const dispatchUserAction = useUserActionDispatchContext();
